refactor: 확인 버튼을 PaymentWidget으로 끌어올려 카드 전체 폭으로 노출
