+++ src/components/PaymentWidget.styles.ts
@@ -7,4 +7,16 @@ export const Wrapper = styled.div`
   border-radius: 10px;
   display: flex;
   flex-direction: column;
+  overflow: hidden;
+`;
+
+export const ConfirmButton = styled.button`
+  width: 100%;
+  height: 52px;
+  background-color: #333333;
+  color: #F3F3F3;
+  border: none;
+  font-size: 16px;
+  font-weight: 500;
+  cursor: pointer;
 `;
