+++ src/components/PaymentWidget.tsx
@@ -1,15 +1,30 @@
+import { useNavigate } from 'react-router-dom';
 import CardInfo from "./cardInfo/CardInfo";
 import CardPreview from "./cardPreview/CardPreview";
-import { Wrapper } from "./PaymentWidget.styles";
+import { Wrapper, ConfirmButton } from "./PaymentWidget.styles";
 import { useCardForm } from "./useCardForm";
+import { useCardStep } from "./cardInfo/useCardStep";
+import { CARD_BRANDS } from "../constants/constants";
 
 export default function PaymentWidget() {
   const cardForm = useCardForm();
+  const step = useCardStep(cardForm);
+  const navigate = useNavigate();
+
+  const handleConfirm = () => {
+    navigate('/complete', {
+      state: {
+        cardNumberFirstSegment: cardForm.cardNumber.value[0],
+        cardBrand: CARD_BRANDS[cardForm.cardBrand.value].label,
+      },
+    });
+  };
 
   return (
     <Wrapper>
       <CardPreview cardForm={cardForm} />
-      <CardInfo cardForm={cardForm} />
+      <CardInfo cardForm={cardForm} step={step} />
+      {step >= 5 && <ConfirmButton onClick={handleConfirm}>확인</ConfirmButton>}
     </Wrapper>
   );
 }
