+++ src/components/cardInfo/CardInfo.stories.tsx
@@ -1,18 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { MemoryRouter } from "react-router-dom";
 import CardInfo from "./CardInfo";
 import { useCardForm } from "../useCardForm";
+import { useCardStep } from "./useCardStep";
 
 const meta: Meta<typeof CardInfo> = {
   title: "Components/CardInfo",
   component: CardInfo,
-  decorators: [
-    (Story) => (
-      <MemoryRouter>
-        <Story />
-      </MemoryRouter>
-    ),
-  ],
 };
 
 export default meta;
@@ -22,6 +15,7 @@ type Story = StoryObj<typeof CardInfo>;
 export const Default: Story = {
   render: () => {
     const cardForm = useCardForm();
-    return <CardInfo cardForm={cardForm} />;
+    const step = useCardStep(cardForm);
+    return <CardInfo cardForm={cardForm} step={step} />;
   },
 };
